Show all blocks of a given type
This is a first step at improving the API. Specifically, this shows how to visit all blocks in a world. Ideally, this logic will be moved into World and allow a callback to be fired on each visited block. Likewise, we should allow the same for regions, chunks, sections?, etc.